Fix #4 by refactoring next id allocation
This commit adds new App object that is responsible for giving the next
free id for new Counters. The next id is always one greater than the
currently greatest id value present in App. As ids are not saved to
localstorage, the id of the same counter may vary between page reloads.
However, the title of the counter will stay the same, and id is only
used to tell counters apart.